CXSPA-10780: pretty file
