+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/my-account/accessibility/tabbing-order.e2e-my-account-v2.cy.ts
@@ -12,47 +12,51 @@ import { standardUser } from '../../../../sample-data/shared-users';
 import { tabbingOrderConfig as config } from './tabbing-order.config';
 import { isolateTests } from '../../../../support/utils/test-isolation';
 
+describe(
+  'Tabbing order - tests do require user to be logged in display model',
+  { testIsolation: false },
+  () => {
+    isolateTests();
 
-describe('Tabbing order - tests do require user to be logged in display model', { testIsolation: false }, () => {
-  isolateTests();
-
-  beforeEach(() => {
-    standardUser.registrationData.email = generateMail(
-      randomString(),
-      true
-    );
-    loginHelper.registerAndLogin(standardUser.registrationData.email, standardUser.registrationData.password);  });
-
-  afterEach(() => {
-    cy.restoreLocalStorage();
-  });
-
-  context('My Account V2 Profile Management ', () => {
-    it('should allow to navigate with tab key display mode (CXSPA-4442)', () => {
-      myAccountV2UserProfileManagementTabbingOrder(
-        config.myAccountV2ProfileDisplay
+    beforeEach(() => {
+      standardUser.registrationData.email = generateMail(randomString(), true);
+      loginHelper.registerAndLogin(
+        standardUser.registrationData.email,
+        standardUser.registrationData.password
       );
     });
 
-    it('should allow to navigate with tab key edit mode (CXSPA-4442)', () => {
-      myAccountV2UserProfileManagementTabbingOrder(
-        config.myAccountV2ProfileEdit,
-        true
-      );
+    afterEach(() => {
+      cy.restoreLocalStorage();
     });
-  });
 
-  context('My Account V2 Email Management', () => {
-    it('should allow to navigate with tab key display mode (CXSPA-4442)', () => {
-      myAccountV2UserEmailManagementTabbingOrder(
-        config.myAccountV2EmailDisplay
-      );
+    context('My Account V2 Profile Management ', () => {
+      it('should allow to navigate with tab key display mode (CXSPA-4442)', () => {
+        myAccountV2UserProfileManagementTabbingOrder(
+          config.myAccountV2ProfileDisplay
+        );
+      });
+
+      it('should allow to navigate with tab key edit mode (CXSPA-4442)', () => {
+        myAccountV2UserProfileManagementTabbingOrder(
+          config.myAccountV2ProfileEdit,
+          true
+        );
+      });
     });
-    it('should allow to navigate with tab key edit mode (CXSPA-4442)', () => {
-      myAccountV2UserEmailManagementTabbingOrder(
-        config.myAccountV2EmailEdit,
-        true
-      );
+
+    context('My Account V2 Email Management', () => {
+      it('should allow to navigate with tab key display mode (CXSPA-4442)', () => {
+        myAccountV2UserEmailManagementTabbingOrder(
+          config.myAccountV2EmailDisplay
+        );
+      });
+      it('should allow to navigate with tab key edit mode (CXSPA-4442)', () => {
+        myAccountV2UserEmailManagementTabbingOrder(
+          config.myAccountV2EmailEdit,
+          true
+        );
+      });
     });
-  });
-});
+  }
+);
