+++ projects/storefrontapp-e2e-cypress/cypress/helpers/my-account-v2/my-account-v2-login-helper.ts
@@ -23,14 +23,16 @@ export function listenForCreateVerificationToken(): string {
 }
 
 export function listenForUserVerficationCodeEmailReceive(
-  customerEmail: string, limitCount: number
+  customerEmail: string,
+  limitCount: number
 ) {
   const mailCCV2Url =
     Cypress.env('MAIL_CCV2_URL') +
     Cypress.env('MAIL_CCV2_PREFIX') +
     '/search?query=' +
     customerEmail +
-    '&limit=' + limitCount;
+    '&limit=' +
+    limitCount;
 
   cy.wait(5000); // allow time for email event handlers
 
@@ -54,12 +56,9 @@ export function listenForTokenAuthenticationRequest(): string {
   return `@${aliasName}`;
 }
 
-export function extractVerficationCode() {
-
-}
+export function extractVerficationCode() {}
 
 export function registerAndLogin(email: string, password: string) {
-
   visitLoginPage();
   cy.get('button.btn-register').click();
   cy.get('cx-otp-register-form form').within(() => {
@@ -68,8 +67,12 @@ export function registerAndLogin(email: string, password: string) {
       .get('div.ng-option')
       .contains('Mr')
       .click();
-    cy.get('[formcontrolname="firstName"]').clear().type(standardUser.registrationData.firstName);
-    cy.get('[formcontrolname="lastName"]').clear().type(standardUser.registrationData.lastName);
+    cy.get('[formcontrolname="firstName"]')
+      .clear()
+      .type(standardUser.registrationData.firstName);
+    cy.get('[formcontrolname="lastName"]')
+      .clear()
+      .type(standardUser.registrationData.lastName);
     cy.get('[formcontrolname="email"]').clear().type(email);
     cy.get('[formcontrolname="termsandconditions"]').click();
     cy.get('button[type=submit]').click();
@@ -77,9 +80,7 @@ export function registerAndLogin(email: string, password: string) {
   listenForUserVerficationCodeEmailReceive(email, 1);
 
   const mailCCV2Url =
-    Cypress.env('MAIL_CCV2_URL') +
-    Cypress.env('MAIL_CCV2_PREFIX') +
-    '/search';
+    Cypress.env('MAIL_CCV2_URL') + Cypress.env('MAIL_CCV2_PREFIX') + '/search';
 
   cy.request({
     method: 'GET',
@@ -87,8 +88,8 @@ export function registerAndLogin(email: string, password: string) {
     qs: {
       query: email,
       limit: 3,
-      sort: 'desc'
-    }
+      sort: 'desc',
+    },
   }).then((response) => {
     const verificationCodeEmailStartText =
       'Please use the following verification code to register in Spartacus Electronics Site: ';
@@ -104,27 +105,24 @@ export function registerAndLogin(email: string, password: string) {
     );
     listenForTokenAuthenticationRequest();
     cy.get('cx-registration-verification-token-form').within(() => {
-      cy.get('[formcontrolname="tokenCode"]')
-        .clear()
-        .type(verificationCode);
+      cy.get('[formcontrolname="tokenCode"]').clear().type(verificationCode);
       cy.get('[formcontrolname="password"]').clear().type(password);
-      cy.get('[formcontrolname="passwordconf"]')
-        .clear()
-        .type(password);
+      cy.get('[formcontrolname="passwordconf"]').clear().type(password);
 
       cy.get('button[type=submit]').click();
       const loginPage = waitForPage('/login', 'getLoginPage');
-      cy.wait(`@${loginPage}`)
-        .its('response.statusCode')
-        .should('eq', 200);
+      cy.wait(`@${loginPage}`).its('response.statusCode').should('eq', 200);
     });
   });
 
   loginWithOTP(email, password, 3);
-
 }
 
-export function loginWithOTP(email: string, password: string, mailMessageLimitCount: number) {
+export function loginWithOTP(
+  email: string,
+  password: string,
+  mailMessageLimitCount: number
+) {
   listenForCreateVerificationToken();
 
   cy.log(`🛒 Logging in user ${email} from the login form`);
@@ -144,31 +142,28 @@ export function loginWithOTP(email: string, password: string, mailMessageLimitCo
   listenForUserVerficationCodeEmailReceive(email, mailMessageLimitCount);
 
   const mailCCV2UrlV2 =
-    Cypress.env('MAIL_CCV2_URL') +
-    Cypress.env('MAIL_CCV2_PREFIX') +
-    '/search';
+    Cypress.env('MAIL_CCV2_URL') + Cypress.env('MAIL_CCV2_PREFIX') + '/search';
 
   cy.request({
     method: 'GET',
     url: mailCCV2UrlV2,
     qs: {
       query: email,
-      limit: mailMessageLimitCount
-    }
+      limit: mailMessageLimitCount,
+    },
   }).then((response) => {
-    const subject =
-      '[Spartacus Electronics Site] Login Verification Code';
+    const subject = '[Spartacus Electronics Site] Login Verification Code';
     const verificationCodeEmailStartText =
       'Please use the following verification code to log in Spartacus Electronics Site: ';
 
     const items = response.body.messages;
-    const sortedMessages = items.sort((a, b) => {
-      const dateA = new Date(a.Created).getTime();
-      const dateB = new Date(b.Created).getTime();
-      return dateB - dateA;
-    }).filter(item =>
-      item.Subject === subject
-    );
+    const sortedMessages = items
+      .sort((a, b) => {
+        const dateA = new Date(a.Created).getTime();
+        const dateB = new Date(b.Created).getTime();
+        return dateB - dateA;
+      })
+      .filter((item) => item.Subject === subject);
 
     const emailBody = sortedMessages[0].Snippet;
 
@@ -184,9 +179,7 @@ export function loginWithOTP(email: string, password: string, mailMessageLimitCo
 
     listenForTokenAuthenticationRequest();
     cy.get('cx-verification-token-form form').within(() => {
-      cy.get('[formcontrolname="tokenCode"]')
-        .clear()
-        .type(verificationCode);
+      cy.get('[formcontrolname="tokenCode"]').clear().type(verificationCode);
       cy.get('button[type=submit]').click();
     });
     cy.wait('@tokenAuthentication')
@@ -194,4 +187,3 @@ export function loginWithOTP(email: string, password: string, mailMessageLimitCo
       .should('eq', 200);
   });
 }
-
